TRANSIMS Version 4.0 December 2008 Release from AECOM Branch 785
SysLib
The field processing for Link_Dir_Files and Link_Data_Files was modified to be more flexible for ArcDelay types of applications.  These files no longer require fields by time period.

The data range class was modified to accept negative values (i.e., directional links) in the range list.

The related file logic was modified to only check for optional files when check data and renumbering flags are set.  It no longer requires that data records exist for these files.

The Num_Records method for SQLITE3 was modified to account for the actual number of records in the table rather than the maximum record number.

A bug was fixed in the link data file class.

TripSum 4.0.12
Additional conditional processing checks were included.  SELECT_PROBABILITY_FILE, SELECT_PROBABILITY_FORMAT, TIME_PERIOD_EQUIVALENCE, and RANDOM_NUMBER_SEEK keys were added to select a sample of trip records based on probabilities specified by origin-destination zone and time period.

LocationData 4.0.9
Error messages were replaced by warning messages when new transit or use fields already exist in the file.

PlanCompare 4.0.20
The plan comparison is now based on total travel time rather than just the drive portion of the trip.  This enables the program to compare transit trips as well as auto trips.

TransimsNet 4.0.20
Logic improvements were made to avoid problems caused by link edits that invalidate pocket lanes and activity locations.  UPDATE_NODE_FILE, UPDATE_LINK_FILE, DELETE_NODE_RANGE, DELETE_LINK_RANGE, DELETE_NODE_FILE, and DELETE_LINK_FILE keys were added to enhance editing flexibility.  Update ranges can now be provided in a file as well as the UPDATE_*_RANGE control key.  Delete ranges and files were provided to remove links or nodes and all of their associated components from the network files. A bug generated by the update/delete logic was fixed.  The link processing logic was enhanced to pass the NOTES field from the input link file to the output link file.  Checks were added to ensure that the delete node list does not includes node numbers required for any of the links written to the output link file.

PlanSelect 4.0.24
SELECT_LINKS_# and SELECT_TRAVEL_MODES keys were added to select highway and transit trips based on links and/or mode codes.  The SELECT_NODES_# behavior was enhanced to mimic the link selection logic that enables transit passengers to be selected based on the driver paths of the transit vehicles.  A bug was fixed related to the mode selection flag to permit mode zero (drive) to be selected by default.

Relocate 4.0.3
New program to update activity locations and parking lots in trip, activity, and vehicle files that are removed or renumbers as a result of TransimsNet (or manual) network edits.  PLAN_FILE, PLAN_FORMAT, NEW_PLAN_FILE, and NEW_PLAN_FORMAT keys were added to implement plan file processing as well.  The logic was also modified to permit the processing of trip and activity files (along with plan and vehicle files) at the same time. Modifications were made to improve the way the household list partitioning is done when trip and activity files are processed simultaneously or in combination with plan files.  A minor change was made in the way the number of households in the output household list file is reported.

Microsimulator 4.0.52
Protection was added in the problem file processing for plans from/to parking lots that are not included in the network.  Conflicts between transit vehicle priority and wait time priority were eliminated.  Warning messages were added to list the lanes approaching an actuated signal that do not have a detector.

LinkSum 4.0.20
A bug was fixed in the link data file interface.

IntControl 4.0.12
Corrections were made to the detector logic for links with no thru lanes.  Detectors are now added to right turn lanes as well to avoid warning messages in the Microsimulator.