SysLib An error message was added to the Plan file class to minimize conflicts between the traveler scaling factor and the scaling factor posted in the Plan definition file.
The traveler scaling factor key was moved from demand service into execution service to avoid problems caused by control methods that read local keys prior to demand keys.  This affected the checking of the scaling factor in the plan definition file (or creating a def file) against a traveler scaling factor that is not yet set.

The Dir List class was split into Dir List and Dir Connect to separate the connection functionality from the link list functionality.

Relocate 4.0.5
The activity location mapping was refined for locations with a positive zone value to skip locations in the new activity location file with zero zone values.  This avoids selecting a transit activity location for a trip activity location.

VISSIMNet 4.0.0
New program to convert a TRANSIMS network to VISSIM XML input format.