SysLib The read activity method within demand service was updated to progress all of the fields in the activity file.
Relocate 4.0.6
Records are now written to the new activity file.

Router 4.0.54
The bug was fixed in positioning the dummy vehicle when the IGNORE_VEHICLE_ID key is used.

NewFormat 4.0.6
Additional logic was added to find locations for activities with no coordinates in the travel survey.  Home locations outside the modeling region were also excluded.

TransitNet 4.0.17
A bug was fixed in the offset calculation for the stop activity location added to the B to A direction of a cross street for a rail station on a link that does not permit walk access.

CheckSurvey 4.0.4
Code was added to insert a notes field in the output survey activity file if the create notes flag is set and the input survey file does not include a notes field.