SysLib The first group and next group methods were added to the link equiv class.
The survey person array class was added.

Progression 4.0.4
GROUP_PERIOD_WEIGHT_FILE key was added to provide weighting factors by time period for each link group.  KEEP_LINK_GROUP_ORDER key was added to manually override the order of the progression settings based on the record order in the link equivalence file.  A bug was fixed in the group length calculation when a link equivalence file is provided.

Microsimulator 4.0.54
The print file now reports the time when the maximum number of vehicles is on the network and the total hours of travel and average travel time for completed vehicle trips.

LinkSum 4.0.21
SELECT_BY_LINK_GROUP key was added to select link using the link equivalence file.

LocationChoice 4.0.4
A bug was fixed initializing the data fields used for the attraction balancing file.

VISSIMNet 4.0.3
Vehicle type and class logic was added.  New network generation logic was added to convert TRANSIMS parking locations to parking access links for VISSIM.  The new network can be output into TRANSIMS format as well as the VISSIM XML format.

PlanCompare 4.0.24
The print file now reports the total hours of travel for the plans selected for output.

ActGen 4.0.34
NEW_HOUSEHOLD_PERSON_COUNT key was added to summarize the number of times a survey household person was replicated.

ActivityPattern 4.0.1
NEW_HOUSEHOLD_PERSON_COUNT key was added to summarize the number of times a survey household person was replicated.

PlanSum 4.0.46
HOUSEHOLD_PERSON_COUNT key was added to summarize the plans with a replication count by household and person.  This is primarily used to generate weighted summary statistics for routed survey activity file.

TripSum 4.0.14
HOUSEHOLD_WEIGHT_FILE was replaced with HOUSEHOLD_PERSON_COUNT to summarize the trips and activities with a replication count by household and person.  This is primarily used to generate weighted summary statistics for survey activity files.