IntControl 4.0.17 The link delay turning movement volumes are now used to create signals with the volume method.  A phasing bug was fixed for intersections with multiple thru lanes sharing the same conflicting thru lane.  The updated logic combines the multiple thru lanes if the approach does not include left or right turns and provides a fully protected phase if turns are permitted.  The thru phase timing weight was also adjusted for situations where leading left-thru phases are generated.  Refinements were also made in assigning detectors to actuated signal movements.  The new logic removes detectors from permitted lefts when protected lefts also exist and removes detectors from thru movements during protected left phases.  Walkway links were ignored by the intersection analysis.  The volume aggregation now is customized for the specific time range specified for each signal in the signalized node file when regenerating signals using the volume method and updating timing plans.
PlanCompare 4.0.25
IGNORE_LOCATION_DIFFERENCES key was added to facilitate compares of traveler trips loaded to different networks.  Additional information about the plan matching process was added to the printout file.  In situations where the complete trip does not match, but the drive legs within the trip do match, a comparison of the drive legs is used to record differences and select plans.  ONLY_COMPARE_DRIVE_LEGS key was added to limit comparisons to matching drive legs.  A bug was fixed in the path dump report.  The path and time dump reports can now select records based on difference parameters greater than 100 percent.  The percent time difference parameter can now be entered as zero to enable selections limited to path differences.  The aggregate difference statistics are now printed even when the distribution reports are not requested.