Route 4.0.55 Logic was added to reset the vehicle location for each person within a household to avoid vehicle access problems caused by another driver not returning the vehicle home at the end of the day.
PlanCompare 4.0.26
A bug was fixed in the path-only comparison process.  The path comparison process was changed to more accurately account for small differences in both the input and compare plans rather than just the compare plan.  The permitted range for maximum percent selected was changed to 0.1 to 100.0.

LinkSum 4.0.23
The performance measure calculations for the previous link delay file were made identical to the delay processing calculations within demand service.

ArcNet 4.0.21
SELECT_TIME_PERIOD key was added to define a time range for selecting time-based network attributes for signals, lane-use, parking, tolls, turns, and transit routes.

SubareaNet 4.0.10
The highest zone number in the activity location file is rounded up by 10 and used as the initial zone number assigned to the activity locations added to each subarea boundary link.

SubareaPlans 4.0.31
The highest zone number in the activity location file is rounded up by 10 and used as the initial zone number assigned to the activity locations added to each subarea boundary link.