ProblemSum 4.0.4 NEW_PROBLEM_TIME_FILE, NEW_PROBLEM_TIME_FORMAT, NEW_PROBLEM_LINK_FILE, and NEW_PROBLEM_LINK_FORMAT keys were added to create files summarizing different problem types by time of day and link by time of day.
PlanSum 4.0.47
Logic was added to correctly process skims and trip tables for plan files that do not start and end at activity locations.  The process link file is read to create a map between parking lots (and transit stops) and connected activity locations to retrieve a zone number for the parking lot.