SysLib The performance data structure and demand service reading process now include the option to summarize three additional fields: congested VMT, congestion VHT, and congestion percentage of the time period.  A Congested_Ratio method was added to set the time ratio that defines congested conditions.  These fields can also be selected for output in a link data or link direction file using the key words CONGESTED_VMT, CONGESTED_VHT, and CONGESTED_TIME.
LinkSum 4.0.25
CONGESTED_TIME_RATIO key was added to activate the congestion performance measures for the detailed performance report and the output link data or link direction files.  If the time period data has a travel time ratio greater than the specified value (default is 3.0), the VMT, VHT, and time period is added to corresponding congestion performance measure.  The new link data and link direction field names that output this data are CONGESTED_VMT, CONGESTED_VHT, and CONGESTED_TIME.  The time values are output as a percentage of the total time period.