SysLib The Link Delay file time period processing within the Demand Service class was updated to properly detected a Version3 format based on the *.def file rather than the default or user specified format key.
DynusTNet 4.0.2
The control.dat file phasing and sign records were modified to include fixed column formatting to work in DynusT as well as DynaSmart.  The sign logic was updated to set two-way and four-way stop codes based on the presence of uncontrolled approaches rather than the number of stop signs.  The zone numbers were re-indexed to eliminate all zones that do not contain a node.  A new file called zone_map.dat was created to equate the TRANSIMS zone numbers to the DynusT zone numbers.

DynusTPlan 4.0.2
The zone numbers were re-indexed to eliminate all zones that do not contain a node for compatibility with DynusTNet.