SysLib AB and BA methods in Link Data file class were changed from private to public.
TripSum 4.0.16
The SUMMARY_LENGTH_INCREMENT key was activated to control the summary increments for trip length reports.

Validate 4.0.16
A warning message was added to the Link Data file processing when the field time ranges do not fit within the summary time periods and increments specified in the control keys.

PlanSum 4.0.48
The trip length skim was corrected in situations where the summary time periods do not span the full travel time of the trip.