SysLib Projection_Flag, Set_Projection, and Get_Projection methods were added to the network service.  If a projection is defined, the default node, zone, and shape processing uses the projection method to converts the XY coordinates.
The projection service classes were reorganized to minimize the data overhead for projection functions in network service.

A warning message was added to the link delay processing method when the period size and the start time are not aligned on an even period boundary.

TransimsNet 4.0.22
Coordinate projection keys were added to convert the input node, zone, and shape coordinate data to meters.  KEEP_ZONE_CONNECTORS_AS_LOCALS key was added to include internal zone connectors in the TRANSIMS network as local streets.

DynusTNet 4.0.3
The closest zone calculation was improved to assign a zone number to a node even if the closest zone centroid is extremely far away.

DynusTPlan 4.0.3
The closest zone calculation was improved to assign a zone number to a node even if the closest zone centroid is extremely far away.