SysLib Driver_Process was modified to avoid a potential null pointer problem when the path link-node combinations are node included in the network.
Emissions 4.0.15
REPORT_TRAVEL_UNITS and REPORT_EMISSIONS_UNITS keys were added to control the units in the output emissions reports.  Options for Travel units include MILES, KILOMETERS, METERS, and FEET.  Options for Emissions units include KILOGRAMS and GRAMS.  The default is KILOMETERS and KILOGRAMS.  The VMT and VHT generated by the speed smoothing process was normalized to equal the VMT and VHT generated by speed bin seconds and cell sizes.

ArcNet 4.0.22
Errors in the transit driver path for a given route were changed to warning messages to continue processing the routes that include valid data.

Microsimulator 4.0.63
Logic was added to constrain the offset and length of lane use restrictions to the cells between the setbacks at both ends of the links.  This eliminates problems caused by lane connectivity / lane use conflicts when entering a new link.