SysLib Min_Value and Max_Value methods were added to the Lookup_Data class.
The read processing methods for link delay and performance files was modified to apply minimum and maximum time range values to the time period calculations.

An additional calling option was added to the vehicle shape function to enable the procedure to draw vehicles with more than two shape points.  

LinkSum 4.0.27
NEW_PERFORMANCE_DATA_FILE and NEW_PERFORMANCE_DATA_FORMAT keys were added to enable the user to write the Performance Detail Report information to a file that can more easily be imported to spreadsheet analysis software.

ArcSnapshot 4.0.13
The link summary file output was modified to include in addition to the total vehicles and persons for the selected vehicle types, the fields containing the number of vehicles and persons by vehicle type as well.  The vehicle shapes for vehicles that are more than two cells long was modified to draw an articulated vehicle based on the shape of the link.

Router 4.0.58
LINK_DELAY_START_TIME and LINK_DELAY_END_TIME keys were added to control the way time periods in the input link delay file are used for travel time calculations. By default, volume and travel time after midnight are added to the corresponding early morning time periods and trips on the network after midnight use these early morning travel times.  The new keys can be used to extend the time period processing beyond 24 hours.