SysLib The lane use processing logic now checks the lane number against the link file.
The problem file format was modified to include a new ROUTE field to store the transit route number for the problem.  This data was previously stored in the LINK field.

ArcSnapshot 4.0.14
ADD_PASSENGER_CIRCLE_SIZE and ADD_PASSENGER_SQUARE_SIZE keys were added to insert two new fields into the output shape files (“RADIUS” and “SQUARE”) that convert the passengers to the radius of circles whose area is equal to passengers and the size of a square whose area is equal to passengers.

PlanCompare 4.0.31
SELECT_BY_RELATIVE_GAP key was added to select plans based on the percent difference in their input and compare travel times.  This key is only active is SELECT_BY_PERCENT_DIFFERENCE key is false.

Microsimulator 4.0.65
A bug was fixed in the way diagonal lane changes are processed for multi-cell vehicles.  Start and end events for transit boardings and alightings are now processed.  In situations where two lanes are blocked using lane use restrictions, the parking access logic will attempt to load the vehicle to the first unrestricted lane.  The problem file processing was modified to save the transit route number in the new ROUTE field rather than the LINK field.  The transit departure time problems now store the link ID of the origin transit stop in the LINK / DIR fields.  A bug was fixed in the way transit boardings were loaded onto the first stop of a route.  

ProblemSum 4.0.5
The logic was modified to detect if a network link and node files are provided before attempting to read the link data.