SysLib The vehicle type data class was modified to permit loading rates greater than 10 boardings per second (i.e., loading rates less than 0.1).
Time_Stamp, Minute, Second, Day, and Time functions were added to the user program functions to convert a time string to a time stamp and extract time components.  The script compiler was enhanced to permit the same text string to be used for function names and variable names.

Transit Capacity and Transit Dwell problem messages were added to record instances when a transit traveler is unable to board a bus or train due to vehicle capacity or maximum wait time constraints.

PlanSum 4.0.50
A bug was fixed in the offset to the destination parking lot used for travel time updates of link-based plan files.

Router 4.0.59
A minor improvement was made to the way minimum wait time was implement for transit path building. 

TripPrep 4.0.0
New program for manipulating trip and activity files using selection variables or user program scripts.

Microsimulator 4.0.68
The logic was updated to enable transit boarding rates greater than 10 per second. The procedure for setting the maximum wait time for a transit vehicle was modified to handle early and late arrivals more realistically.  Transit Capacity and Transit Dwell problem messages were added to record instances when a transit traveler is unable to board a bus or train due to vehicle capacity or maximum wait time constraints.