SysLib The lane use processing logic now checks the lane number against the link file. The Line data class was expanded to include an impedance penalty for selected runs and stops of a route.
PlanSum 4.0.51
The transit load factor is now calculated when a vehicle type file is provided.

Router 4.0.60
TRANSIT_PENALTY_FILE key was added to input impedance penalties for combinations of stop, route, and/or run to adjust the likelihood that a stop will be included in a path based on the capacity constraints of the transit service.