SysLib The Event file and output service and the Problem file were expanded to include the trip leg number.
Reschedule 4.0.9
RUN_START_TIME_FILE key was added to update the stop schedules for a route based on the difference between the departure times for the first stop of each run.  This is designed to synchronize the run start times between transit networks.

Microsimulator 4.0.69
The Event output service, traveler plan data, and problem messages were expanded to include the trip leg number.

Router 4.0.61
A bug was fixed in the stop impedance penalty.  The logic was also modified to make the stop number optional (i.e., the penalty can be applied to all stops on a route and/or run).