SysLib Data range class was modified to accept zero as a range.
EventSum 4.0.6
The Event data and sort index and the travel time file were expanded to include the trip leg.  SELECT_TRIP_MODES and SELECT_PLAN_MODES keys were added to filter the trip and event records that are processed.  

Reschedule 4.0.10
TRANSIT_SERVICE_FACTOR key was added to increase or decrease the number of runs on selected routes or modes during selected time periods.

Microsimulator 4.0.70
The person code for transit driver events was changed to zero to avoid conflicts with household persons.