Reschedule 4.0.11 Corrected the time offset for the service factoring.  Added capability to factor multiple time ranges in the same application. TIME_PERIOD_METHOD key was added to control the way routes are associated with time period ranges.
EventSum 4.0.7
SELECT_HOUSEHOLD_RANGE key was added limit processing to a range of household IDs.  COMBINE_EVENT_LEGS key was added to enable plan leg events to be collapsed into a single trip record.  Logic was added to convert plan modes to trip modes and then use the trip mode to select records to include in the output travel time file.

Microsimulator 4.0.72
Process and save walk legs in event file.  A bug was fixed in the way the first walk leg of multiple trips for a given traveler was output to the event file.

LineSum 4.0.11
Added logic to summarize the transit ridership in link group reports.