Router 4.0.63 The processing checks were modified to permit the plan update to be executed without a household list or an input trip and/or activity file.  The printout now includes a message about the number of updates that were made.  PRINT_UPDATE_WARNINGS key was added to control the warning messages sent to the print file.
PlanCompare 4.0.33
SELECT_TRAVELERS key was added to restrict the plan comparison to a range of traveler IDs.

SubareaPlans 4.0.32
The fare zone number assigned to the new subarea boundary stop at the end of the line was corrected.

PlanSelect 4.0.31
A bug was fixed in the link and node selection procedures for transit legs.