TRANSIMS Version 4.0.6.1 -- AECOM Branch 1005
PlanPrep 4.0.12
A bug caused by a recent change to the plan file data structure to support generalized cost updates and plan re-skimming was fixed.  This bug impacts subarea merges of HOV trips.

LinkSum 4.0.28
A bug was fixed for link group selection criteria that include the positive and negative orientation of a link in the same link group.

Microsimulator 4.0.75
Changes were made to the way maximum dwell times are implemented at transit stops.  If the vehicle is unable to leave the stop after the maximum dwell time is reached, it will no longer open its doors to permit additional passengers to board.

SubareaNet 4.0.11
A bug was fixed in the schedule calculation for the starting boundary stop on transit routes.

SubareaPlans 4.0.33
A bug was fixed in the schedule calculation for the starting boundary stop on transit routes.
PlanCompare 4.0.35
NEW_START_DIFFERENCE_FILE and NEW_END_DIFFERENCE_FILE keys were added to summarize the plan time or cost values and differences by origin location and/or destination location (activity location or parking lot).  A COUNT field was added to the start and end difference files to store the number of observations included in the cumulative times/costs and differences.  INDEX_BY_COMPARE_LOCATION key was added to control which location is used for indexing the difference files when the IGNORE_LOCATION_DIFFERENCES key is true.