Transims 4.2 Updates
SysLib
The descriptive text for pocket lanes types was improved and the number of output pocket lane records was corrected.

Links with no-sign records are no longer written to the output sign file.

Db_Base Copy_Fields and Replicate_Fields were changed from references to pointers to simplify coding in several places.  The binary string read function was modified to remove trailing null characters.  A nest_flag was added to convert nested fields into a flat file.

The timing codes assigned to version 4.0 signalized node records were modified to remove gaps in the number sequence.

The timing plan output service was modified to sort the nested phase numbers for each signal.

Logic was added to the Dtime class to process time data in minutes.

Metric to English unit switches were moved into the Db_Base class.

Static_Service was reconfigured to exclude the Execution_Service pointer and include the functions from Utility.hpp as methods.  Utility.hpp and Utility.cpp were deleted. The pointer to Executation_Service was made a global external pointer.  Message_Service now derives from Static_Service.  Default_Format, Time_Format, and Metric_Flag were moved to Static_Service

Control_Service was expanded to include methods to check the existence of a key and return keys of integer, double, and Dtime types.  These new methods include the capability of reading and translating the units of measures from control key strings.  Default key values and units are read from the “help” field within the control key structure.  Similar logic is used convert the data to the appropriate output units for writing to the printout file with unit descriptors.  A method was added to output key error message and perform range checks on length keys with appropriate unit conversions if the length is out of range.

A bug in the Buffer Size method was fixed.

Z coordinates were added to the Zone file and data structures.

The ArcView file class was redesigned to derive from a Points (XYZ) vector and use private XY_Points and Doubles vectors for the XY and Z/M coordinates needed for native file input and output.

ArcNet 4.2.0
Converted from 4.0 to 4.2 file formats.  SELECT_TIME key was added to output network attributes for a specified time of day.