TRANSIMS Version 4.2 SysLib Lanes_Code and Link_Dir data structures were added to standardize bit packing logic.  Several data classes were modified to use these new services.
A bug was fixed in the link shape drawing method.

The logic for converting Version 4.0 detector offsets was corrected.

ArcNet 4.2.0
The arcview path and filename logic was modified to strip existing paths and extensions from the input network file names and replace them with the arcview directory and extensions.  Turn Penalty file was added and drawn using lane connection shapes.  Lane use drawing was enhanced to account for pocket lane lengths and offsets.  Detector and transit route files were added.