TRANSIMS Version 4.0
SysLib
A bug was fixed in the user program lookup table initialization process.

LocationData 4.0.11
A bug was fixed in the subzone allocation distance calculation. 

ConvertTours 4.0.2
NEW_DIURNAL_FILE key was added to create a file that reports the smoothed diurnal distribution, the target trip distribution, and the actual trip distribution for each group and diurnal data column and each minute of the day.