TRANSIMS Version 4.0
SysLib
The time difference class was modified to store the data as a double to avoid integer overflow problems that report negative absolute values.

ConvertTours 4.0.5
Work_Code was added to the user program return data to identify a tour as a work tour or an at-work tour.  Code 1 is used to identify a work tour and 2 is used to identify an at-work tour. The work location and vehicle ID are now mapped to specific household persons.  The return time for trips and the start time for at-work tours are now adjusted based on the end time of the prior trip.  Location and time schedule warning messages are now included in the count of non-allocated trips.  Warning messages are now generated when the origin or destination is zero as a result of a previous location problem. STOP_WEIGHT_FIELD_# was added to specify the field in the activity location file used to locate intermediate stops.  The stop location logic was modified to give preferences to stops that minimize the distance to the origin and destination location.