TRANSIMS 4.0
LocationData 4.0.12
NET_ZONE_FILE, SUBZONE_ZONE_FACTOR_FILE and SUBZONE_ZONE_FIELD_* keys were added to enable a data field in the zone file to be distributed to activity locations within the zone based on the distance weighted values of a data field in a subzone file.  The zone factor file enables a single subzone record to be distributed proportionally to multiple zones for calculating the activity location weights within each zone.

PlanCompare 4.0.36
NEW_TRAVELER_MATCH_FILE key was added to write traveler records that do and do not match.

ConvertTours 4.0.8
The logic was modified to sort the trips made by each household person by start time to integrate subtours with primary tours in order to avoid vehicle access problems in the Router. The logic was added to ensure that trips with end time constraints don’t start earlier than the end of the previous trip.