TRANSIMS Version 4.0 -- March 2010 Release
SysLib
A bug was fixed in the user program lookup table initialization process.

The time difference class was modified to store the data as a double to avoid integer overflow problems that report negative absolute values.

TransimsNet 4.0.26
POCKET_TO_FACILITY_BY_AREA_TYPE key was added to control the minimum facility type by area type a pocket lane can turn into when deciding to add a pocket lane to a link.

TPPlusNet 4.0.3
The program now detects if the input TP-Plus link or node file is an ArcView shape file and accordingly reads the shape points for export to a NEW_SHAPE_TABLE.  The speed-capacity lookup table logic was expanded to include two digit facility type, area type, and lane values and to interpret the value modification flags (+, -, *) in front of the capacity and speed data.  FACILITY_INDEX_FIELD and AREA_TYPE_INDEX_FIELD keys were added to permit the user to read a field from the input link file as the index into the speed-capacity lookup table.  CONVERT_MPH_TO_MPS key was added to force a conversion of the speed data in the speed-capacity lookup table from MPH to MPS.  Warning messages are also generated if the speed values are less than 0.1 MPS.

Microsimulator 4.0.76
A change was made to eliminate link access problem messages when moving between links.

LocationData 4.0.12
Logic was added to allocate subzone data to activity locations proportional to the distance to the subzone centroid.  These calculations are activated by setting MAX_SUBZONE_DISTANCE to zero.  A bug was fixed in the subzone allocation distance calculation. NET_ZONE_FILE, SUBZONE_ZONE_FACTOR_FILE and SUBZONE_ZONE_FIELD_* keys were added to enable a data field in the zone file to be distributed to activity locations within the zone based on the distance weighted values of a data field in a subzone file.  The zone factor file enables a single subzone record to be distributed proportionally to multiple zones for calculating the activity location weights within each zone.

PlanSum 4.0.53
Logic was added to protect against the trip length calculation errors when the origin or destination parking location has an offset of zero.

PlanCompare 4.0.36
NEW_TRAVELER_MATCH_FILE key was added to write traveler records that do and do not match.

ConvertTours 4.0.8
New program to convert tour-based travel files to TRANSIMS trip, household, population, and vehicle files.  MODE_OUT and MODE_IN fields were added to the script output file.  Multiple tours from the same household start at the same home location.  Trip numbers are now person-based rather than tour-based.  The same vehicle number is used for multiple tours of a given person within the household.  NEW_DIURNAL_FILE key was added to create a file that reports the smoothed diurnal distribution, the target trip distribution, and the actual trip distribution for each group and diurnal data column and each minute of the day. Work_Code was added to the user program return data to identify a tour as a work tour or an at-work tour.  Code 1 is used to identify a work tour and 2 is used to identify an at-work tour. The work location and vehicle ID are now mapped to specific household persons.  The return time for trips and the start time for at-work tours are now adjusted based on the end time of the prior trip.  Location and time schedule warning messages are now included in the count of non-allocated trips.  Warning messages are now generated when the origin or destination is zero as a result of a previous location problem. STOP_WEIGHT_FIELD_# was added to specify the field in the activity location file used to locate intermediate stops.  The stop location logic was modified to give preferences to stops that minimize the distance to the origin and destination location. The distance weight logic for intermediate stops was modified to use the furthest location within the target zone at the maximum distance reference.  The logic was modified to sort the trips made by each household person by start time to integrate subtours with primary tours in order to avoid vehicle access problems in the Router. The logic was added to ensure that trips with end time constraints don’t start earlier than the end of the previous trip.