TRANSIMS Version 4.0
Router 4.0.64
NEW_PATH_ATTRIBUTE_FILE key was added to write detailed summary information about the components of a traveler’s trip including the time walking, first waiting, transfer waiting, in transit, and in driving, the number of transfers, the cost, and the total impedance and impedance penalties added to transit boardings and transfers.

PlanSum 4.0.54
The capacity calculation used in volume-delay equations was modified to permit a single time period of more than four hours.

LinkSum 4.0.29
The capacity calculation was modified to account for time increments of more than four hours.