TRANSIMS Version 4.0.7 -- Candidate Release Software Changes between 4.0.6 and 4.0.7
SysLib
A bug was fixed in the user program lookup table initialization process.

The time difference class was modified to store the data as a double to avoid integer overflow problems that report negative absolute values.

The generalized cost distribution data were enlarged by a factor of six to accommodate a greater range of difference values than are typically experienced with travel time differences.  The increased distribution range addressed problems in calculating 85th percentiles less than negative $60 or greater than $90 dollars.

The time/cost difference class was enhanced to summarize and write the average walk time, drive time, transit time, and travel cost differences for each increment in the generalized cost or total travel time difference distribution.

Router 4.0.64
The re-skimming logic was corrected for transit plans with the waiting time assigned to the access leg.  NEW_PATH_ATTRIBUTE_FILE key was added to write detailed summary information about the components of a traveler’s trip including the time walking, first waiting, transfer waiting, in transit, and driving, the number of transfers, the cost, and the total impedance and impedance penalties added to transit boardings and transfers.

Microsimulator 4.0.76
A change was made to eliminate link access problem messages when moving between links.

PlanPrep 4.0.13
SELECT_TRAVELER_RANGE key was added to copy a range of traveler plans to the output file.

PlanSum 4.0.54
Logic was added to protect against the trip length calculation errors when the origin or destination parking location has an offset of zero.  The capacity calculation used in volume-delay equations was modified to permit a single time period of more than four hours.

PlanCompare 4.0.41
NEW_TRAVELER_MATCH_FILE key was added to write traveler records that do and do not match.  INPUT_TRIP_FILE, COMPARE_TRIP_FILE, and NEW_INCOMPLETE_TRIP_FILE keys were added to ensure that the match process considers only completed trips in the comparison results.  Time and cost data for incomplete trips can be written to the incomplete trip file.  SELECT_TRIP_MODES and NEW_MODE_DISTRIBUTION_FILE keys were added to report mode-related comparisons.  If trip files are provided, the user can select the trip mode types to be processed.  The mode distribution file provides the average walk time, drive time, transit time, and travel cost differences associated with each increment in the generalized cost or travel time difference distribution.  HIGHEST_MATCH_DIFFERENCE and LOWEST_MATCH_DIFFERENCE keys were added to enable the user to limit the statistical comparison of plans to a reasonable range of values.  Activity location data was added to compare coordinate locations for trip origins and destinations.  NEW_MODE_DISTRIBUTION_DATA_FILE key was added to write the average values for walk, drive, and transit times and cost for the base and compare alternatives.

ConvertTours 4.0.9
New program to convert tour-based travel files to TRANSIMS trip, household, population, and vehicle files.  MODE_OUT and MODE_IN fields were added to the script output file.  Multiple tours from the same household start at the same home location.  Trip numbers are now person-based rather than tour-based.  The same vehicle number is used for multiple tours of a given person within the household.  NEW_DIURNAL_FILE key was added to create a file that reports the smoothed diurnal distribution, the target trip distribution, and the actual trip distribution for each group and diurnal data column and each minute of the day. Work_Code was added to the user program return data to identify a tour as a work tour or an at-work tour.  Code 1 is used to identify a work tour and 2 is used to identify an at-work tour. The work location and vehicle ID are now mapped to specific household persons.  The return time for trips and the start time for at-work tours are now adjusted based on the end time of the prior trip.  Location and time schedule warning messages are now included in the count of non-allocated trips.  Warning messages are now generated when the origin or destination is zero as a result of a previous location problem. STOP_WEIGHT_FIELD_# was added to specify the field in the activity location file used to locate intermediate stops.  The stop location logic was modified to give preferences to stops that minimize the distance to the origin and destination location. The distance weight logic for intermediate stops was modified to use the furthest location within the target zone at the maximum distance reference.  The logic was modified to sort the trips made by each household person by start time to integrate subtours with primary tours in order to avoid vehicle access problems in the Router. The logic was added to ensure that trips with end time constraints don’t start earlier than the end of the previous trip.  The tour processing logic was modified to sort the input tour file records within a household by person, start time, and end time to avoid problems caused by trips and subtours that are out of time order.  

ConvertTrips 4.0.20
A bug was fixed in the travel time calculation when distance weights are disabled and skims are not provided.

LocationData 4.0.12
Logic was added to allocate subzone data to activity locations proportional to the distance to the subzone centroid.  These calculations are activated by setting MAX_SUBZONE_DISTANCE to zero.  A bug was fixed in the subzone allocation distance calculation. NET_ZONE_FILE, SUBZONE_ZONE_FACTOR_FILE and SUBZONE_ZONE_FIELD_* keys were added to enable a data field in the zone file to be distributed to activity locations within the zone based on the distance weighted values of a data field in a subzone file.  The zone factor file enables a single subzone record to be distributed proportionally to multiple zones for calculating the activity location weights within each zone.

EventSum 4.0.9
NEW_DIFFERENCE_FILE key was added to output the travel time differences for each traveler.

LinkSum 4.0.29
The capacity calculation was modified to account for time increments of more than four hours.

Relocate 4.0.13
ZONE_RELOCATE_FILE, RANOM_NUMBER_SEED, and NEW_CHANGE_FILE keys were added to randomly distribution trips within, to or from a list of zones to new activity locations within the zone.  The vehicle locations are updated accordingly.

TripPrep 4.0.1
SELECT_HOUSEHOLDS key was added to copy a range of household trips to the output file.

TransimsNet 4.0.26
POCKET_TO_FACILITY_BY_AREA_TYPE key was added to control the minimum facility type by area type a pocket lane can turn into when deciding to add a pocket lane to a link.

TPPlusNet 4.0.3
The program now detects if the input TP-Plus link or node file is an ArcView shape file and accordingly reads the shape points for export to a NEW_SHAPE_TABLE.  The speed-capacity lookup table logic was expanded to include two digit facility type, area type, and lane values and to interpret the value modification flags (+, -, *) in front of the capacity and speed data.  FACILITY_INDEX_FIELD and AREA_TYPE_INDEX_FIELD keys were added to permit the user to read a field from the input link file as the index into the speed-capacity lookup table.  CONVERT_MPH_TO_MPS key was added to force a conversion of the speed data in the speed-capacity lookup table from MPH to MPS.  Warning messages are also generated if the speed values are less than 0.1 MPS.